Fix operators catalog downloader.
The previous way to get operators catalog pages wasn't retrieving
all the pages. It can be tested manually in the browser that there are
many more pages than the calculated ones.

As we cannot rely on the total value to calculate the number of
operator pages, the only option is reading until we found a page without any
operator.