Fix for platform-alteration-base-image test case.
The current podman diff --json output won't match the expected regex, as
the output won't come as in the original UT, which added indentation
before the folder and it was expected by the regex.

The actual output currently comes as a single line, like this:
{"changed":["/root","/usr","/usr/lib"],"added":["/root/.bash_history","/usr/lib/blablah"]}

I've changed the implementation to unmarshall the output in order to be more
clear and verbose in the log/claim file when the test case fails.

Also, added lib64 folders to be tested.