Fetch tool workaround for container pages download.
There's an performance issue in the Pyxis catalog which is preventing
the TNF's fetch tool to make the first query for the containers catalog,
which tries to get the page_size and the number of certified containers.

This workaround ignores both the total and the page_size and just
queries page by page until one of them returns fewer entries than the
desired number, which has been hardcoded to 100.

Signed-off-by: Gonzalo Reyero Ferreras <greyerof@redhat.com>