Remove fetch tool to use OCT instead (#677)
* Remove fetch tool to use OCT instead

The fetch tool as well as the DB for containers, helm charts and
operatos are removed from the CNF repo. The OCT tool should be used instead
to build the DB that CNF consumes. Also, as of this change there is no
need for a Github workflow to maintain this internal DB.