Get rid of 'oc' binary dependency. (#701)
We used the oc command to detect whether the suite was running against
an OCP cluster or not (kind/minikube) in the cnf-run-suites.sh script.
If no OCP was detected by the 'oc' command, the TNF_NON_OCP_CLUSTER
env var was exported to be "true".

As we already used the go-client to get the OCP version, I just
changed the code to use that information to avoid the need for the
external TNF_NON_OCP_CLUSTER.

The getOpenshiftVersion() has been modified to return err only in case
the client's Get() operation failed with error different than
"ErrIsNotFound" or the field "versions.Name" is not the expected one in
the ClusterOperator object returned. Both cases will lead to stop the
execution as it doesn't make any sense to run any TC in that context.