Refactor for the kernel taints test case (#754)
The whitelist in the previous implementation was for taints related to
out-of-tree modules. This commit tries to make the whitelist to work for
any bit, including the RedHat's ones.

First, I get the list of modules that have any taint letter, reading
the /sys/module/<name>/taint files. When the module is responsible for a
taint bit, that module will have a string letters associated to the
taints. If any of those modules doesn't appear in the whitelist, the
test fails.

Finally, we need to make sure there's no other taint bits active,
because some of them can be active for other reasons not related to
kernel modules (e.g. bit 11).

Also, updated description for taint bit 31.