platform-alteration-base-image test case workaround. (#1005)
The previous implementation of this test cases fails in OCP 4.12.7+ due
to podman installed on those OCP worker nodes (v4.2.0) using an older
version of the containers/storage lib than the one cri-o uses on those
nodes. As a result, "podman diff" won't work because it tries to read a
file "containers.json" that has been renamed to
"volatile-containers.json" in latest OCP builds.

This workaround is based on a (compatible) precompiled podman binary
existing in a /root/podman folder of the debug pods. Currently, v4.4.0
is the version that can be built with UBI 8 that can read the new
"volatile-containers.json" inside worker nodes and fix the "podman diff"
function.

In order to avoid copying the binary into the node, I prefer mounting
the containing folder into node's /tmp/tnf-podman so the program can be
run with /tmp/tnf-podman/podman.

The new implementation of the test case does the following:
1. Creates the temp folder /host/tmp/tnf-podman
2. Mounts debug pod's /root/podman into /host/tmp/tnf-podman
3. Calls /chroot /host /tmp/tnf-podman diff <container-id>
4. Unmounts /host/tmp/tnf-podman
5. Removes /host/tmp/tnf-podman