Avoid crashing when discoverying CR scales.
The previous code was crashing because it was trying to get the Scale
object of each CR found, without considering whether the CRD spec was
configured to have a scale subresource. In those cases, using the Get()
of the dynamic client to retrieve the Scale corresponding to those CRs
was returning a NotFound err.

More changes:
- The scale objects won't be searched for cluster-wide CRDs.
- I did some refactoring to the functions related to searching
  the scales and creating the custom ScaleObject lists.
- File names were also changed to pure snake case.