Allow CNF Cert Suite app's go-clients to work in a k8s pod.
This change will allow the go-clients to be initialized properly in case
the app is running inside a pod in a k8s/OCP cluster.

The method rest.InClusterConfig() reads the pod's SA tokens/creentials
and returns a rest.Config struct that can be used to initialize all the
go-clients.

If that method returns an error, it assumes it's running outside a
cluster, using the provided filenames to create the rest.Config.