Improved helper function that waits for podsets to be ready. (#1091)
The helper function WaitForAllPodsetReady was taking too long in case
there were too many deployments+statefulsets to check. In that case, as
it was waiting sequentially for each of them to time out (7mins), the
test case lifecycle-pod-recreation could take hours to finish in the
worst case scenario (none of the podsets are ready).

I've refactored it to check the status of every podset in every pass and
wait 15secs before checking again the not-ready ones. There's no
per-podset timeout, just a global timeout that needs to be passed for
all of them to be ready.