Changed annotation used to parse multus networks.
We were parsing the plural "networks-status", but that annotation was a
mistake introduced by multus devs long ago, and they left them there for
backwards compatibilty issues. It can be seen in any pod that both
"networks-status" and "network-status" annotations have the same value.
See:
https://github.com/k8snetworkplumbingwg/multus-cni/issues/468#issuecomment-607977200

In OCP 4.13, that annotation won't be added/used to pods any more, in
favor of the singular version "network-status".