Fixed unstable performance suite test cases.
Not sure why, but "ps" seems not to be able to handle the URG signal,
returning an error status code. As per some comments in github, looks
like the golang runtime might have sent the socket related URG signal.
See:
https://github.com/golang/go/issues/37942
https://stackoverflow.com/questions/72568024/go-preempt-can-exit-out-of-a-loop

The ps program shows this output:
[command.go: 72] stderr: Signal 23 (URG) caught by ps (3.3.15).

Apparently, ps works well, but it just returns with an error status
code due to the received SIGURG.

A retry mechanism doesn't work well here: in my tests, I saw the ps
command failing up to 5 times in a row, so I think it's better trapping
the signal with the bash built-in command "trap". This is the new
command sent to the node's shell to get all the processes:

const command = "trap \"\" SIGURG ; ps -e -o pidns,pid,args"

Also, I refactored the way the pids from the container namespaces are
retrieved, as the nsenter used by the previous code is not actually
needed. Regular "lsns" and "ps" commands running in the node's debug
container (with chroot /root) are good enough to get the final lists of
pids from processes that are running in the same namespace as the
container's one. In fact, as per the stdout logs of the nsenter
commands, the "ps" that was running with nsenter was showing all the
processes running in that node.

After the refactor, the schedulling UT TestGetProcessCPUScheduling was
failing because it was mocking a function that was not used inside the
target function. I tried fixing it, but then I decided to remove it as
I realized it was not covering more code than the existing UT
TestExistsSchedulingPolicyAndPriority.