CLI tool subcommand to show failed test cases from a claim file.
Added a new subcommand to the tnf cli tool that parses a claim file and
shows the failed test cases in a more readable way. By default, the
output is plain text, but a 'json' output mode can be selected in order
to get a json representation of those failed test cases per test suite.

Also, the subcommand has a flag --testsuites that can be used to reduce
the output to the tests suites included in the comma separated list.

Examples:
1. Show all the failed test cases from all test suites:
./tnf claim show failures --claim path/to/claim.json

2. Show all the failed test cases from all tests suites in json output.
./tnf claim show failures --claim path/to/claim.json --output json

3. Show all the failed test cases from the lifecycle and access-control
   suites.
./tnf claim show failures --claim path/to/claim.json --testsuites
"lifecycle,access-control"