Avoid crashing when the failed test case does not use report objects.
The code didn't generate the failures report when one of the failed test
cases was not using the report objects, which made the json.Unmarshal
function to fail.

This change makes the tool to avoid that, printing the raw failureReason
string in case it's not possible to unmarshal it to a list of
NonCompliantObjects.

To avoid polluting the output with the unmarshaling errors from these
test cases, stderr is recommended to be sent to a file a file or to
dev/null:
./tnf claim show failures --claim claim.json -o json 2>/dev/null
{
  "testSuites": [
    {
      "name": "access-control",
      "failures": [
        {
          "name": "access-control-sys-admin-capability-check",
          "description": "Ensures that containers do not use SYS_ADMIN capability",
          "failureReason": "pod xxx ns yyy container zzz uses SYS_ADMIN"
        }
      ]
    }
  ]
}