Claim comparison tool improvement: results table.
This is the first of several commits/PRs to improve the claim comparison
tool. The goal is to compare most (if not all) of the claim sections of
both claim files so the output shows the differences in a table based
format.

This change is aimed to show just the differences in the results of each
test case.

Two tables will show the results differences from both files, as well as
a summary table showing passed/skipped/failed numbers.

Sample output:
...
TEST CASE NAME                                              CLAIM-1   CLAIM-2
access-control-net-admin-capability-check                   failed    passed
access-control-pod-automount-service-account-token          passed    failed
access-control-sys-nice-realtime-capability                 passed    failed
...

STATUS         # in CLAIM-1        # in CLAIM-2
passed         22                  21
skipped        62                  62
failed         3                   4