Avoid zipping junit xml file in diagnostic mode.
An ERROR message was appearing at the end when the test suite was
launched in diagnostic mode (no labels used), because the archiver
package was trying to read a non-existent xml junit file.

ERROR  [Jul 25 19:36:43.837][suite_test.go: 226] Failed to compress
results artifacts: failed to get file info from
cnf-certification-tests_junit.xml: stat
cnf-certification-tests_junit.xml: no such file or directory

The error itself was not making the program to return any error code,
thus it was not detected in github ci.

As that file is never created in diagnostic mode, this change just
avoids adding it to the list of files to zip at the end. Also, the
program will return with error code if there's some error while creating
the results+html zip file.