Add tnf.Log() function to allow log level selection
It avoids the common pattern of calling tnf.ClaimFilePrint() followed or
preceded by a logrus function to print a log with a different level than
TRACE. It also avoids log duplication when doing this.

The issue is fixed in a couple of places as an example, not with the
intent of being exhaustive.