cmd/tnf: new tool to generate the CNF config file with user input (#1506)
* cmd/tnf: new tool to generate the CNF config file with user input

The tool presents a set of menus that allow the user to configure the
different sections of the CNF configuration. The explanation of the
options is provided along with some examples and other useful
information. The result is a YAML file that can be used when running the
CNF Certification Suite.

Usage ./tnf generate config