performance: avoid failing test case if a process cannot be found (#1564)
* performance: avoid failing test case if a process cannot be found

It applies to rt-apps-no-exec-probes.

The reason a process cannot be found after it has been properly
retrieved previously could be that the process has finished. So, in this
case it seems more adequate to just ignore it than to produce an error.
If the error is of another kind the container is added to the non
compliant list but the processing for that container continues instead
of breaking, to have more info on the processes.

The commit also changes the verification order in that test case, making
the faster check of exec probes at the beginning, skipping the slower
part in which the processes are list.

Finally, the compliant containers are added to its list with the reason
for compliance.