Transition certification suite to ginkgoless.
Non-backward compatibility: the tc/check
"affiliated-certification-helm-version" will be skipped if no helm
chart releases were discovered. Also, in case we discovered some chart
releases and helm is v3, every chart release will be added as a compliant
report object.

Also, added the expansion of the "args..." in the ClaimFilePrintf call
to Logf.