log: add new log package that uses slog as global logger (#1663)
* log: add new log package that uses slog as global logger

The log package provides top-level funtions for creating log records in
four different log levels as well as some other helper functions to
set up the logging parameters and create special log handlers.

Two slog handlers are created: the MultiHandler that allows distributing
the logs to several io.Writer objects and the CustomHandler to produce
the log output in a specifc format.

The checksdb.Check type is also modified to add two new fields: a logger
and a log archive where the check's log will be stored. The check's
logger has the two handlers described above chained, so that the log
output is written to the log file and to the check's log archive.

A log file has been created with a default name cnf-test-log.

Only a couple of log lines have been produces using this new logger as
an example. The bulk of the work for converting the old log lines so
that they use the new package will be done in a separate change.