Avoid loading preflight checks in web server mode.
This is a temporary workaround to allow running the app in web server
mode when no kubeconfig file/env var has been provided.

For preflight, the clientsholder needs to be fully and correctly
initialized in order to run the checks, otherwise it won't have any
TestEnvironment to get the containers/operators under test.

For web server mode, there's no need to have a valid clientsholder yet,
as the kubeconfig file is provided via web form, and it can be a
different one on each run. The clientsholder is then created with the
clientsholder.GetNewClientsHolder().

This workaround should probably be removed/refactored when PR #1684 is
merged.