Non graceful abort for unrecoverable/critial failures inside checkFn.
The check.Abort() has been modified to panic with a message wrapped in
a checksdb type. The deferred function in runCheck() will do nothing in
case that especial (own) panic is detected, and no other check of that
group will run.

This way we stop the goroutine for group.RunChecks() as soon as possible
to avoid race conditions with the main goroutine, which will call
group.OnAbort() and will register all the remaining checks as skipped,
using the reason from the msg channel.