log: remove loghelper package (#1721)
The functions in the loghelper package are not necessary anymore with
the new log package as the check multilogger is able to log into the log
file as well as into each check's log archive, which can be later used to
fill the log section in the claim file.

Helper functions for the checks receive now the logger instance of the
check, so that logs can be written inside those functions but tagged with
the check's name and recorded in that check's log archive. For this, a
new Logger type has been defined, as well as a set of methods to be used
by the logger instance.

With the help of the new Logger type a small refactor of the log package
is included in this change. Now the "log/slog" library is only used
inside this package, isolating the rest of the code from any other
logging libray than our internal log package.