CLI output improvements: check's elapsed time and last log trace.
The aim of this change is adding a bit of visual responsiveness to the
current cli output of the ginkoless branch when the checks run.

For each running check it will show the elapsed time and the last log
line emitted using check.Log() func.

Once the check finishes, the cli line will also be rewritten but only
the result will be left.

E.g.: while the check is running:
[ RUNNING ] lifecycle-pod-recreation (4s)   INFO  [Dec 18 07:50:14.778]
[suite.go: 642] [lifecycle-pod-recreation] Draining and Cordoning node
kind-worker3:

When it's finished:
[ PASS ] lifecycle-pod-recreation

Also, if the check was skipped, aborted/errored, the reason will also
appear so the user can quickly have an idea of what happened to that
check.

E.g:
[ SKIP ] performance-max-resources-exec-probes  (Not matching labels)

This per-running check's extra info is shown only in case the program
runs in a tty/ptty mode (manual run through cnf-run-suites.sh).