Add new tnf command to check the test results
The command compares the test results of a CNFCERT execution using the
log file with those provided in a YAML template to check that they match.
If not, a table showing the expected vs actual results will be printed
on stdout.

To generate the test results YAML template from a log file:

* tnf check results --generate-template

To check the test results against a reference template:

* tnf check results --template <expected_results.yaml> --log-file
  <cnf-certsuite.log>