Fix for platform-alteration-base-image test case. (#1861)
The original workaround to use a custom podman for this test case
stopped working when we switched our debug pod's image from ubi8 to ubi9
due to libc missing in RHEL8.x based nodes (OCP 4.12.z and lower). This
pre-built podman binary was used in this tc no matter the OCP version of
the cluster.

This commit will make the test case to use our ubi8-based podman only if
the test suite is checking CNFs on OCPs 4.12.z or lower. For newer
versions of OCP, the podman that's presinstalled in the nodes will be
used, as it seems to work well again.

We should remove this ugly workaround when 4.12 reaches EOL.

The fix to build podman with ubi8 again is already merged in the partner repo
here:
https://github.com/test-network-function/cnf-certification-test-partner/pull/394

The UT that test the folder creation/mount issues of podman need OCP ver
4.12 or lower to run.