autodiscover: modify Istio Service Mesh detection (#1896)
The Istio Service Mesh detection required a two step process:
  1) Check for the existence of the "istio-system" namespace.
  2) Check for the presence of a CR from the Istio Operator called
     "installed-state".

But it has been reported that when Istio is installed using the Helm
chart method this CR is not created. So, the step 2 above has been
replaced for checking the presence of the "istiod" Deployment, which is
in charge of running the Istio discovery service.