Avoid unnecessary rebuild of the test environment (#1919)
* Avoid unnecessary rebuild of the test environment

The logs show that the test environment is built twice before starting running the test cases, which should not be necessary. However, disabling this rebuild of the test environment lead to an error while marshaling the TestConfiguration to JSON format. This PR refactors the way the Preflight test results are processed in order to produce an output of the test environment adequate for JSON marshaling without the need to rebuild it.

The problem came with the direct addition of the struct plibRuntime.Results into our Container struct. This former struct contains an interface that prevents the struct to be properly transformed to JSON. To solve this issue, another struct called PreflightResultsDB that only contains the information needed has been used instead.

The test environment rebuild hid this problem because the Preflight results were cleared in the process, allowing the JSON marshaling to work correctly. The results could be retrieved later (even after being cleared) due to the fact that the old results were stored in the closure that contains the test function (defined in WithCheckFn()).