Fix for test case operator-install-status-no-privileges. (#1665)
* Fix for test case operator-install-status-no-privileges.

With this fix, the test case fails when resourceName's rules have been
found in the CSV's clusterPermissions field and the operator was not
cluster-wide installed.

Also, minor refactor to use a helper pointer to the env.Operators[i].

* Adapted to new logger after rebasing.

* Modified test case to detect SCC rules in CSV's clusterPermissions.

The original motivation for this test case was checking whether the
operator wasn't creating any SA with RBAC rules related to SCC.

The implementation was wrong as it was (only) checking the resourceNames
list to be empty to pass. But empty list means that any resourceName is
allowed (as per k8s docs).

Since it's impossible to predict all the "privileged" SCCs that might
have been created in the cluster apart from the default ones, this new
implementation is another "canary in the mine" check so we can flag
operators that try to use/create cluster's SCCs.

Operators will need to justify the SCC accesses in their SAs.

* Updated compliant reason when no clusterPermissions found.

Also, minor refactor to remove the need for the helper var
securityResourceFound.