docs: CNF Certification rename to Red Hat Best Practices for Kubernetes
The references to "CNF Certification" or "CNF Certification Suite" have
been changed to "Red Hat Best Practices Test Suite for Kubernetes" or a
simplified version of this name, such as "Test Suite", when the meaning
is clear from the context.

Also, to broaden the scope of the potential targets of the Test Suite
the work "CNF" has been replaced for "workload" where applicable.