cmd/tnf: add subcommand to run the test suite
Use ./tnf run <flags> to run the test suite. See ./tnf run -h for more
info.

The original way of running the test suite using the
"cnf-certification-test" binary and the bash scripts is still available
and used in the GitHub CI. This change just add the possibility of
running the test suite with the "tnf" command.