Parse flags after env vars to override them if required
To support the transition from the standard test suite binary to the
test suite run as a "certsuite run" subcommand the precedence of flags
needs to be higher than that of the env vars, as flags will be the main
mechanism for passing options to the test suite.

As no new flags will be added to the old flag library, it's necessary to
leave the env vars but override them when the test suite it run with the
"certsuite" command.