Added Test Case Implementation subsection to docs site.
Added a new subsection called "Test Case Implementation" in order to
describe the implementation details of some test cases that might
confuse the users as the CATALOG descripcion might not be so clear or
not cover all the scenarios.

Flow/sequence/state charts should be used as much as possible as they
usually describe processes better than words.

Added implementation details for:
- lifecycle-deployment/statefulset-scaling test cases.
- lifecycle-crd-scaling

Also, I've added the name of the test cases that are considered
"intrusive" in the "runtime-env.md" section of the related env var.