Dockerfile: do not build the legacy binary
The container image is updated to avoid building the old binary and
copying stuff that it is not required anymore.

The "certsuite" binary is installed in "/usr/local/bin" so that it can
be directly called.

The GitHub workflows have been updated accordingly.

An update in DCI and QE will be necessary as well.