Delete legacy code and avoid using environment variables
As the old binary is no longer used , its "main" as well as its flags
have been deleted.

The struct "TestParameters" that was used to parse the environment
variables is now used for storing the "certsuite run" command's flags
information.

The code that was used for supporting the two running methods has also
been deleted and the Makefile has been updated.