Fix for the false positive of access-control's capabilities tcs.
The forbidden capabilities test cases in access-control test suite are
giving false positives. The code was stringifying the whole
"capabilities" field and searching for the capability in the resulting
string, which led to false positives when the capability was set in the
"drop" list, as we've seen in a partner.

The new code checks whether it was explicitly set in the Add list
directly, matching by capability name or by "ALL".