Cluster-wide operators support improvements
This commit enables the discovery of cluster-wide operator pods
(controllers), but also singleNamespace/multiNamespace installation
methods where the targetNamespaces were set to be different than the
installation namespace. The code will look for the controller pod in the
installation namespace always. Operand pods can always be tested using
normal namespace or label discovery methods, but a best-effort to find
them in the configured namespaces will be implemented in a follow-up PR.

As a reminder:
- installation namespace: namespace where both the subscription and the
  operator group where created. This is where the operator's controller
  (aka operator pod) is deployed by OLM.
- targetNamespaces: namespaced where the controller will watch for CRs.
  The installation namespace can appear in this list but it's not
  mandatory.

Also, the test cases of the operator test suite related to controller
pods have been moved to the access-control test suite, as those
requirements also apply to any workload. For that reason, the discovered
controller pods have also been added to the normal env.testPods so all
the checks can be performed on them.

To distinguish between securityContext.runAsUser and
securityContext.runAsNonRoot, the original test case
access-control-security-context-non-root-user-check has been renamed to
access-control-security-context-non-root-user-id-check