Operand pods autodiscovery
This change adds an autodiscovery mechanism for operand pods that
"belong" to CSVs/Operators under test. This is a separate and
complementary way to the by-label/by-namespace already existing
discovery modes.

Usually, operands are deployed by operators by means of a CR. Ideally,
operators designers should also flag that CR as owner of those operand
pods that are deployed in the cluster.

The autodiscovery mechanism for operand pods creates a list of CRDs
owned by the CSVs under test. Then, it iterates over every pod found in
the configured namespaces in order to get whether its owner is a CR of
any of those CRD types. If so, it's added to the list of pods under test
if it wasn't already there.