CSVs autodiscovery fix. (#2653)
PR #2479 fixed the autodiscovery of operator's pods and also added them to
the pods-under-test list. The problem is that it was also adding every
cluster-wide operator's controller pods to that list.

With this change, the operator (controller's pod) must be running in one of the
configured/test namespaces in order to consider that CSV as part of the
operators under test.