Skip access-control-sys-nice-realtime-capability if there's no nodes using realtime kernel.
There's no point in running this check if there isn't any node using
realtime kernel, so I added a custom skip function for it.

Anyway, if nodes with realtime kernel exist, it's still possible/compliant
to deploy workload pods in non-realtime kernel nodes, irrespective of
the sys_nice cap. For pods deployed in realtime kernel nodes, it's
mandatory for this cap to be set at container level.