Fix to disable/enable pre-main self-hosted workflows.
I removed the "disable-self-hosted" label from PR #2763 but the pre-main
self-hosted workflow jobs didn't run.

As suggested in github docs, the way to trigger the workflow when a
label is added removed is using pull_request's activity types.

We did this in the do-not-merge workflow file and it works well there.

See
https://docs.github.com/en/actions/writing-workflows/choosing-when-your-workflow-runs/events-that-trigger-workflows#label

"If you want to run your workflow when a label is added to or removed
from an issue, pull request, or discussion, use the labeled or unlabeled
activity types for the issues, pull_request, pull_request_target, or
discussion events instead."